requirements.txt
beautifulsoup4>=4.12
lxml>=5.0
pillow>=10.0
pypdf>=3.0
pyyaml>=6.0
